deployed

+++ src/pages/api/menu/index.ts
@@ -9,17 +9,20 @@ const menu = async (req: NextApiRequest, res: NextApiResponse) => {
   if (!session) return res.status(500).send("Unauto Rize");
   const method = req.method;
   if (method === "POST") {
-    const { name, price, menuCategoryId } = req.body;
+    const { name, assetUrl, price, menuCategoryId } = req.body;
     const isValid = name && price !== undefined && menuCategoryId.length > 0;
     if (!isValid) return res.status(404).send("Missing Requeire");
-    const newMenu = await prisma.menu.create({ data: { name, price } });
+
+    const newMenu = await prisma.menu.create({
+      data: { name, price, assetUrl },
+    });
 
     const menuCategoryIds: { menuId: number; menuCategoryId: number }[] =
       menuCategoryId.map((item: number) => ({
         menuId: newMenu.id,
         menuCategoryId: item,
       }));
-    const menuCategoryMenu = await prisma.$transaction(
+    const newMenuCategoryMenu = await prisma.$transaction(
       menuCategoryIds.map((item) =>
         prisma.menuCategoryMenu.create({
           data: { menuCategoryId: item.menuCategoryId, menuId: item.menuId },
@@ -27,14 +30,25 @@ const menu = async (req: NextApiRequest, res: NextApiResponse) => {
       )
     );
 
-    return res.status(200).json({ newMenu, menuCategoryMenu });
+    return res.status(200).json({ newMenu, newMenuCategoryMenu });
   } else if (method === "PUT") {
-    const { id, name, price, menuCategoryId } = req.body;
+    const {
+      id,
+      name,
+      price,
+      menuCategoryId,
+      locationId,
+      assetUrl,
+      isAvailable,
+    } = req.body;
     const isValid =
       id && name && price !== undefined && menuCategoryId.length > 0;
     if (!isValid) return res.status(404).send("Missing Required Datas");
+    const exit = await prisma.menu.findFirst({ where: { id } });
+    if (!exit) return res.status(404).send("Missing Required Datas");
+
     const updateMenus = await prisma.menu.update({
-      data: { name, price },
+      data: { name, price, assetUrl },
       where: { id },
     });
     await prisma.menuCategoryMenu.deleteMany({ where: { menuId: id } });
@@ -51,11 +65,38 @@ const menu = async (req: NextApiRequest, res: NextApiResponse) => {
         })
       )
     );
-    return res.status(200).json({ updateMenus, updateMenuCategoryMenu });
+    if (locationId !== undefined) {
+      if (isAvailable === false) {
+        const exit = await prisma.disableLocationMenu.findFirst({
+          where: { locationId, menuId: id },
+        });
+        if (!exit) {
+          await prisma.disableLocationMenu.create({
+            data: { menuId: id, locationId },
+          });
+        }
+      } else {
+        await prisma.disableLocationMenu.deleteMany({
+          where: { menuId: id, locationId },
+        });
+      }
+    }
+    const disableLocationMenus = await prisma.disableLocationMenu.findMany({
+      where: { menuId: id },
+    });
+
+    return res
+      .status(200)
+      .json({ updateMenus, updateMenuCategoryMenu, disableLocationMenus });
   } else if (method === "DELETE") {
     const menuId = Number(req.query.id);
     const menus = await prisma.menu.findFirst({ where: { id: menuId } });
     if (!menus) return res.status(404).send("Data Not Found");
+    await prisma.menuAddonCategory.updateMany({
+      data: { isArchived: true },
+      where: { menuId },
+    });
+
     await prisma.menu.update({
       data: { isArchived: true },
       where: { id: menuId },
